refactor(apps): split workspace app commands by concern
Break the former 2.5k-line apps command module into focused config, detection, launch, model, platform, and test modules while preserving the existing Tauri command surface.

Move the Windows icon extraction PowerShell/C# payload into an included script asset and route macOS, Linux, Windows, and fallback icon/toolbox behavior through a shared platform facade.

Keep existing regression coverage intact for app config parsing, detection ordering, JetBrains defaults, icon data URLs, and platform-specific helpers.
