+++ src-tauri/src/commands/apps.rs
@@ -1,1081 +1,22 @@
-use std::path::{Path, PathBuf};
+mod config;
+mod detection;
+mod launch;
+mod model;
+mod platform;
 
-// `data_url_from_bytes` (the only consumer of these symbols) is gated
-// to macOS + Linux — macOS reads icons out of `.app` bundles, Linux
-// reads them from XDG icon roots — so the import follows.
-#[cfg(any(target_os = "macos", target_os = "linux"))]
-use base64::{Engine as _, engine::general_purpose};
-use serde::{Deserialize, Serialize};
+#[cfg(test)]
+mod tests;
+
 use tauri::State;
 
 use crate::state::AppState;
-#[cfg(target_os = "windows")]
-use claudette::process::CommandWindowExt as _;
 
-const DEFAULT_APPS_JSON: &str = include_str!("../../default-apps.json");
-pub(crate) const DEFAULT_TERMINAL_APP_SETTING_KEY: &str = "default_terminal_app_id";
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AppCategory {
-    Editor,
-    FileManager,
-    Terminal,
-    Ide,
-}
-
-/// Entry in the user-editable apps.json config.
-#[derive(Debug, Clone, Deserialize)]
-pub struct AppEntry {
-    pub id: String,
-    pub name: String,
-    pub category: AppCategory,
-    #[serde(default)]
-    pub bin_names: Vec<String>,
-    #[serde(default)]
-    #[cfg_attr(not(target_os = "macos"), allow(dead_code))]
-    pub mac_app_names: Vec<String>,
-    /// `.exe` filenames to walk up to from `detected_path` for icon
-    /// extraction on Windows. The npm-shim layouts used by VS Code,
-    /// Cursor, etc. put a no-extension bash shim or `.cmd` wrapper in
-    /// PATH while the real `.exe` (the one with embedded icon
-    /// resources) sits one or more directories above. Setting this to
-    /// e.g. `["Code.exe"]` lets Windows builds resolve VS Code's
-    /// actual binary for `ExtractAssociatedIcon`. Ignored on other
-    /// platforms.
-    #[serde(default)]
-    #[cfg_attr(not(target_os = "windows"), allow(dead_code))]
-    pub windows_exe_names: Vec<String>,
-    /// AppxPackage name (or prefix — `Get-AppxPackage -Name "<value>*"`)
-    /// for UWP/MSIX-packaged apps whose real binary sits inside
-    /// `%PROGRAMFILES%\WindowsApps` and is reached through a 0-byte
-    /// execution alias on PATH (e.g. Windows Terminal's `wt.exe`).
-    /// `IShellItemImageFactory` follows the alias to the console
-    /// subsystem and returns a generic glyph; the real vendor logo
-    /// lives in the package's `Assets/*Logo*.png` files. Setting this
-    /// triggers a manifest-aware lookup before the regular
-    /// `windows_exe_names` walk-up. Ignored on other platforms.
-    #[serde(default)]
-    #[cfg_attr(not(target_os = "windows"), allow(dead_code))]
-    pub windows_appx_package: String,
-    pub open_args: Vec<String>,
-    #[serde(default)]
-    pub needs_terminal: bool,
-}
-
-/// The apps.json file structure.
-#[derive(Debug, Clone, Deserialize)]
-pub struct AppsConfig {
-    pub apps: Vec<AppEntry>,
-}
-
-/// App that passed detection (returned to frontend).
-#[derive(Debug, Clone, Serialize)]
-pub struct DetectedApp {
-    pub id: String,
-    pub name: String,
-    pub category: AppCategory,
-    /// The resolved binary path or .app bundle path.
-    pub detected_path: String,
-    /// A platform-resolved application icon as a browser-renderable data URL.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub icon_data_url: Option<String>,
-}
-
-pub(crate) fn select_workspace_terminal_app_id(
-    detected_apps: &[DetectedApp],
-    preferred_app_id: Option<&str>,
-) -> Option<String> {
-    let terminals = detected_apps
-        .iter()
-        .filter(|app| app.category == AppCategory::Terminal);
-
-    if let Some(preferred_app_id) = preferred_app_id.and_then(|id| {
-        let trimmed = id.trim();
-        (!trimmed.is_empty()).then_some(trimmed)
-    }) {
-        if let Some(app) = terminals.clone().find(|app| app.id == preferred_app_id) {
-            return Some(app.id.clone());
-        }
-    }
-
-    terminals.map(|app| app.id.clone()).next()
-}
-
-// ---------------------------------------------------------------------------
-// Config loading
-// ---------------------------------------------------------------------------
-
-/// Resolve the path to the user's apps.json config file. Honors
-/// `$CLAUDETTE_HOME` via [`claudette::path::claudette_home`].
-fn apps_config_path() -> Option<PathBuf> {
-    Some(claudette::path::claudette_home().join("apps.json"))
-}
-
-/// Merge missing entries from `embedded` into `user` without
-/// overwriting any field the user has explicitly set. Two cases:
-///
-/// - **New entries:** any embedded app whose `id` is absent from
-///   `user` is appended verbatim. This lets a Claudette upgrade
-///   ship new editor / terminal entries (e.g. Windows Terminal,
-///   PowerShell 7) without forcing the user to delete their
-///   `apps.json`.
-/// - **Backfill new optional fields:** if the user's entry has an
-///   empty `mac_app_names` / `windows_exe_names`, copy from the
-///   matching embedded entry. These are never user-empty by intent
-///   (the field either matters for the platform or doesn't), and
-///   skipping them would silently break icon extraction for users
-///   whose `apps.json` predates the field being added.
-///
-/// Anything the user has actually customized — `bin_names`,
-/// `open_args`, `name`, `category`, `needs_terminal` — is left
-/// untouched.
-fn merge_missing_default_entries(mut user: AppsConfig, embedded: AppsConfig) -> AppsConfig {
-    use std::collections::HashSet;
-    let user_ids: HashSet<String> = user.apps.iter().map(|a| a.id.clone()).collect();
-
-    for app in user.apps.iter_mut() {
-        let Some(default_entry) = embedded.apps.iter().find(|d| d.id == app.id) else {
-            continue;
-        };
-        if app.mac_app_names.is_empty() && !default_entry.mac_app_names.is_empty() {
-            app.mac_app_names = default_entry.mac_app_names.clone();
-        }
-        if app.windows_exe_names.is_empty() && !default_entry.windows_exe_names.is_empty() {
-            app.windows_exe_names = default_entry.windows_exe_names.clone();
-        }
-        if app.windows_appx_package.is_empty() && !default_entry.windows_appx_package.is_empty() {
-            app.windows_appx_package = default_entry.windows_appx_package.clone();
-        }
-    }
-
-    for default_entry in embedded.apps {
-        if !user_ids.contains(&default_entry.id) {
-            user.apps.push(default_entry);
-        }
-    }
-
-    user
-}
-
-/// Load and parse apps.json from the given path.
-/// If the file doesn't exist, write the embedded default and return it.
-/// If the file is malformed, log a warning and return the embedded default.
-/// If the file parses, merge in any missing entries / fields from the
-/// embedded default so upgrades surface new apps without the user
-/// having to delete their config.
-fn load_apps_config_from(path: &Path) -> AppsConfig {
-    let embedded: AppsConfig =
-        serde_json::from_str(DEFAULT_APPS_JSON).expect("embedded default-apps.json must be valid");
-
-    if path.exists() {
-        match std::fs::read_to_string(path) {
-            Ok(content) => match serde_json::from_str::<AppsConfig>(&content) {
-                Ok(config) => return merge_missing_default_entries(config, embedded),
-                Err(e) => tracing::warn!(
-                    target: "claudette::apps",
-                    path = %path.display(),
-                    error = %e,
-                    "failed to parse apps config"
-                ),
-            },
-            Err(e) => tracing::warn!(
-                target: "claudette::apps",
-                path = %path.display(),
-                error = %e,
-                "failed to read apps config"
-            ),
-        }
-    } else {
-        // Write the default file for the user to discover and customize.
-        if let Some(parent) = path.parent() {
-            let _ = std::fs::create_dir_all(parent);
-        }
-        if let Err(e) = std::fs::write(path, DEFAULT_APPS_JSON) {
-            tracing::warn!(
-                target: "claudette::apps",
-                path = %path.display(),
-                error = %e,
-                "failed to write default apps config"
-            );
-        }
-    }
-    // Fallback: the embedded default always parses.
-    embedded
-}
-
-/// Public entry point — resolves ~/.claudette/apps.json and loads it.
-fn load_apps_config() -> AppsConfig {
-    match apps_config_path() {
-        Some(path) => load_apps_config_from(&path),
-        None => serde_json::from_str(DEFAULT_APPS_JSON)
-            .expect("embedded default-apps.json must be valid"),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Detection logic
-// ---------------------------------------------------------------------------
-
-/// Well-known PATH prefixes that macOS GUI apps may not inherit.
-const EXTRA_PATH_DIRS: &[&str] = &["/opt/homebrew/bin", "/usr/local/bin", "/usr/local/sbin"];
-
-#[cfg(target_os = "macos")]
-fn jetbrains_toolbox_script_dirs(home: Option<&Path>) -> Vec<PathBuf> {
-    home.map(|home| {
-        home.join("Library")
-            .join("Application Support")
-            .join("JetBrains")
-            .join("Toolbox")
-            .join("scripts")
-    })
-    .into_iter()
-    .collect()
-}
-
-#[cfg(target_os = "linux")]
-fn jetbrains_toolbox_script_dirs(home: Option<&Path>) -> Vec<PathBuf> {
-    home.map(|home| {
-        home.join(".local")
-            .join("share")
-            .join("JetBrains")
-            .join("Toolbox")
-            .join("scripts")
-    })
-    .into_iter()
-    .collect()
-}
-
-#[cfg(windows)]
-fn jetbrains_toolbox_script_dirs(_home: Option<&Path>) -> Vec<PathBuf> {
-    dirs::data_local_dir()
-        .map(|dir| dir.join("JetBrains").join("Toolbox").join("scripts"))
-        .into_iter()
-        .collect()
-}
-
-#[cfg(not(any(target_os = "macos", target_os = "linux", windows)))]
-fn jetbrains_toolbox_script_dirs(_home: Option<&Path>) -> Vec<PathBuf> {
-    Vec::new()
-}
-
-/// Build the list of directories to scan for binaries.
-fn build_path_dirs() -> Vec<PathBuf> {
-    let mut dirs = Vec::new();
-
-    for dir in EXTRA_PATH_DIRS {
-        dirs.push(PathBuf::from(dir));
-    }
-    let home_dir = dirs::home_dir();
-    if let Some(home) = home_dir.as_ref() {
-        dirs.push(home.join(".local/bin"));
-    }
-    dirs.extend(jetbrains_toolbox_script_dirs(home_dir.as_deref()));
-
-    if let Some(path_var) = std::env::var_os("PATH") {
-        for dir in std::env::split_paths(&path_var) {
-            dirs.push(dir);
-        }
-    }
-
-    // Deduplicate while preserving order.
-    let mut seen = std::collections::HashSet::new();
-    dirs.retain(|d| seen.insert(d.clone()));
-    dirs
-}
-
-/// Extensions tried in order on Windows, mirroring PATHEXT semantics
-/// closely enough for app detection. `.exe` is probed first so a real
-/// executable wins over Bash-style shims (e.g. VS Code's `bin/code`,
-/// a no-ext shell script that won't run via `CreateProcess` and isn't
-/// `ExtractAssociatedIcon`-friendly). The empty string keeps the legacy
-/// behavior of accepting bare-name matches as a last resort, which Unix
-/// configs entered into Windows PATH may rely on.
-#[cfg(windows)]
-const WINDOWS_BIN_EXTS: &[&str] = &[".exe", ".cmd", ".bat", ""];
-
-/// Check whether `name` exists as an executable in any of `path_dirs`.
-/// Returns the full path to the first match, or `None`.
-fn find_binary(name: &str, path_dirs: &[PathBuf]) -> Option<PathBuf> {
-    for dir in path_dirs {
-        // On Windows, probe each PATHEXT-style extension before moving
-        // to the next directory so a `code.cmd` in dir A wins over a
-        // bare-name `code` farther down PATH.
-        #[cfg(windows)]
-        {
-            for ext in WINDOWS_BIN_EXTS {
-                let candidate = dir.join(format!("{name}{ext}"));
-                let Ok(meta) = std::fs::metadata(&candidate) else {
-                    continue;
-                };
-                if !meta.is_file() {
-                    continue;
-                }
-                return Some(candidate);
-            }
-            continue;
-        }
-        #[cfg(not(windows))]
-        {
-            let candidate = dir.join(name);
-            let Ok(meta) = std::fs::metadata(&candidate) else {
-                continue;
-            };
-            if !meta.is_file() {
-                continue;
-            }
-            // On Unix, verify the executable bit is set.
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if meta.permissions().mode() & 0o111 == 0 {
-                    continue;
-                }
-            }
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-/// Check whether a .app bundle exists in /Applications (macOS only).
-#[cfg(target_os = "macos")]
-fn find_mac_app(app_name: &str) -> Option<PathBuf> {
-    if app_name == "__always__" {
-        // Sentinel: always detected on macOS (e.g. Terminal.app).
-        return Some(PathBuf::from("/System/Applications/Utilities/Terminal.app"));
-    }
-    let roots = [
-        PathBuf::from("/Applications"),
-        PathBuf::from("/Applications/Utilities"),
-        PathBuf::from("/System/Applications"),
-        PathBuf::from("/System/Applications/Utilities"),
-        PathBuf::from("/System/Library/CoreServices"),
-    ];
-
-    if let Some(path) = roots
-        .iter()
-        .map(|root| root.join(app_name))
-        .find(|path| path.exists())
-    {
-        return Some(path);
-    }
-
-    dirs::home_dir()
-        .map(|home| home.join("Applications").join(app_name))
-        .filter(|path| path.exists())
-}
-
-// Used by both `mac_icon_from_app_bundle` (PlistBuddy/sips bytes) and
-// the Linux icon walk (`find_icon_file_recursive` →
-// `image_data_url_from_file`) to synthesize data URLs for the
-// workspace-opener UI. Match the gate to both consumers so Windows
-// builds don't trip `dead_code` under -Dwarnings.
-#[cfg(any(target_os = "macos", target_os = "linux"))]
-fn data_url_from_bytes(media_type: &str, bytes: &[u8]) -> String {
-    format!(
-        "data:{media_type};base64,{}",
-        general_purpose::STANDARD.encode(bytes)
-    )
-}
-
-#[cfg(any(target_os = "macos", target_os = "linux"))]
-fn image_data_url_from_file(path: &Path) -> Option<String> {
-    let ext = path
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(str::to_ascii_lowercase)?;
-    let media_type = match ext.as_str() {
-        "png" => "image/png",
-        "jpg" | "jpeg" => "image/jpeg",
-        "svg" => "image/svg+xml",
-        "webp" => "image/webp",
-        _ => return None,
-    };
-    let bytes = std::fs::read(path).ok()?;
-    Some(data_url_from_bytes(media_type, &bytes))
-}
-
-#[cfg(target_os = "macos")]
-fn mac_icon_from_app_bundle(app_path: &Path) -> Option<String> {
-    let info_plist = app_path.join("Contents/Info.plist");
-    let output = claudette::process::std_command("/usr/libexec/PlistBuddy")
-        .args(["-c", "Print :CFBundleIconFile"])
-        .arg(&info_plist)
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-
-    let icon_name = String::from_utf8(output.stdout).ok()?;
-    let icon_name = icon_name.trim();
-    if icon_name.is_empty() {
-        return None;
-    }
-
-    let icon_filename = if Path::new(icon_name).extension().is_some() {
-        icon_name.to_owned()
-    } else {
-        format!("{icon_name}.icns")
-    };
-    let icon_path = app_path.join("Contents/Resources").join(icon_filename);
-    if !icon_path.exists() {
-        return None;
-    }
-
-    let out_dir = std::env::temp_dir().join(format!(
-        "claudette-app-icon-{}-{}",
-        std::process::id(),
-        uuid::Uuid::new_v4()
-    ));
-    std::fs::create_dir_all(&out_dir).ok()?;
-    let out_file = out_dir.join("icon.png");
-
-    let output = claudette::process::std_command("sips")
-        .args(["-s", "format", "png"])
-        .arg(&icon_path)
-        .arg("--out")
-        .arg(&out_file)
-        .output()
-        .ok();
-
-    let icon = output
-        .filter(|output| output.status.success())
-        .and_then(|_| image_data_url_from_file(&out_file));
-
-    let _ = std::fs::remove_dir_all(&out_dir);
-    icon
-}
-
-#[cfg(target_os = "macos")]
-fn app_icon_data_url(entry: &AppEntry, detected_path: &Path) -> Option<String> {
-    entry
-        .mac_app_names
-        .iter()
-        .find_map(|name| find_mac_app(name))
-        .or_else(|| {
-            detected_path
-                .ends_with(".app")
-                .then(|| detected_path.to_path_buf())
-        })
-        .and_then(|app_path| mac_icon_from_app_bundle(&app_path))
-}
-
-#[cfg(target_os = "linux")]
-fn desktop_file_dirs() -> Vec<PathBuf> {
-    let mut dirs = vec![
-        PathBuf::from("/usr/share/applications"),
-        PathBuf::from("/usr/local/share/applications"),
-        PathBuf::from("/var/lib/flatpak/exports/share/applications"),
-    ];
-
-    if let Some(home) = dirs::home_dir() {
-        dirs.push(home.join(".local/share/applications"));
-        dirs.push(home.join(".local/share/flatpak/exports/share/applications"));
-    }
-
-    dirs
-}
-
-#[cfg(target_os = "linux")]
-fn parse_desktop_value(contents: &str, key: &str) -> Option<String> {
-    let prefix = format!("{key}=");
-    contents
-        .lines()
-        .find_map(|line| line.strip_prefix(&prefix))
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-#[cfg(target_os = "linux")]
-fn normalized_match_text(value: &str) -> String {
-    value
-        .chars()
-        .filter(|c| c.is_ascii_alphanumeric())
-        .flat_map(char::to_lowercase)
-        .collect()
-}
-
-#[cfg(target_os = "linux")]
-fn desktop_file_matches(
-    entry: &AppEntry,
-    detected_path: &Path,
-    file_stem: &str,
-    contents: &str,
-) -> bool {
-    let exec = parse_desktop_value(contents, "Exec").unwrap_or_default();
-    let name = parse_desktop_value(contents, "Name").unwrap_or_default();
-    let detected_name = detected_path
-        .file_name()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default();
-
-    let mut needles = vec![entry.id.as_str(), entry.name.as_str(), detected_name];
-    needles.extend(entry.bin_names.iter().map(String::as_str));
-
-    let desktop_key = normalized_match_text(file_stem);
-    let name_key = normalized_match_text(&name);
-    let exec_key = normalized_match_text(&exec);
-
-    needles.iter().any(|needle| {
-        let key = normalized_match_text(needle);
-        !key.is_empty()
-            && (desktop_key.contains(&key) || name_key.contains(&key) || exec_key.contains(&key))
-    })
-}
-
-#[cfg(target_os = "linux")]
-fn find_linux_desktop_icon_name(entry: &AppEntry, detected_path: &Path) -> Option<String> {
-    for dir in desktop_file_dirs() {
-        let Ok(entries) = std::fs::read_dir(dir) else {
-            continue;
-        };
-        for entry_path in entries.filter_map(Result::ok).map(|entry| entry.path()) {
-            if !entry_path
-                .extension()
-                .and_then(|ext| ext.to_str())
-                .is_some_and(|ext| ext.eq_ignore_ascii_case("desktop"))
-            {
-                continue;
-            }
-            let Ok(contents) = std::fs::read_to_string(&entry_path) else {
-                continue;
-            };
-            let file_stem = entry_path
-                .file_stem()
-                .and_then(|value| value.to_str())
-                .unwrap_or_default();
-            if desktop_file_matches(entry, detected_path, file_stem, &contents) {
-                if let Some(icon) = parse_desktop_value(&contents, "Icon") {
-                    return Some(icon);
-                }
-            }
-        }
-    }
-    None
-}
-
-#[cfg(target_os = "linux")]
-fn linux_icon_roots() -> Vec<PathBuf> {
-    let mut roots = vec![
-        PathBuf::from("/usr/share/pixmaps"),
-        PathBuf::from("/usr/share/icons"),
-        PathBuf::from("/usr/local/share/icons"),
-    ];
-    if let Some(home) = dirs::home_dir() {
-        roots.push(home.join(".local/share/icons"));
-        roots.push(home.join(".icons"));
-    }
-    roots
-}
-
-#[cfg(target_os = "linux")]
-fn icon_candidate_score(path: &Path) -> u8 {
-    let text = path.to_string_lossy();
-    if text.contains("/64x64/") {
-        0
-    } else if text.contains("/128x128/") {
-        1
-    } else if text.contains("/256x256/") {
-        2
-    } else if text.contains("/scalable/") {
-        3
-    } else if text.ends_with(".png") {
-        4
-    } else {
-        5
-    }
-}
-
-#[cfg(target_os = "linux")]
-fn find_icon_file_recursive(root: &Path, icon_name: &str, max_depth: usize) -> Option<PathBuf> {
-    let mut stack = vec![(root.to_path_buf(), 0usize)];
-    let mut best: Option<PathBuf> = None;
-    let mut visited = 0usize;
-
-    while let Some((dir, depth)) = stack.pop() {
-        visited += 1;
-        if visited > 12_000 {
-            break;
-        }
-        let Ok(entries) = std::fs::read_dir(dir) else {
-            continue;
-        };
-        for path in entries.filter_map(Result::ok).map(|entry| entry.path()) {
-            if path.is_dir() {
-                if depth < max_depth {
-                    stack.push((path, depth + 1));
-                }
-                continue;
-            }
-
-            let Some(stem) = path.file_stem().and_then(|value| value.to_str()) else {
-                continue;
-            };
-            if stem != icon_name {
-                continue;
-            }
-            if image_data_url_from_file(&path).is_none() {
-                continue;
-            }
-
-            let replace = best
-                .as_ref()
-                .is_none_or(|current| icon_candidate_score(&path) < icon_candidate_score(current));
-            if replace {
-                best = Some(path);
-            }
-        }
-    }
-
-    best
-}
-
-#[cfg(target_os = "linux")]
-fn linux_icon_file_from_name(icon_name: &str) -> Option<PathBuf> {
-    let icon_path = PathBuf::from(icon_name);
-    if icon_path.is_absolute() && icon_path.exists() {
-        return Some(icon_path);
-    }
-
-    let direct_names =
-        ["png", "svg", "jpg", "jpeg", "webp"].map(|ext| format!("{icon_name}.{ext}"));
-    for root in linux_icon_roots() {
-        for direct_name in &direct_names {
-            let direct = root.join(direct_name);
-            if direct.exists() {
-                return Some(direct);
-            }
-        }
-        if let Some(path) = find_icon_file_recursive(&root, icon_name, 6) {
-            return Some(path);
-        }
-    }
-
-    None
-}
-
-#[cfg(target_os = "linux")]
-fn app_icon_data_url(entry: &AppEntry, detected_path: &Path) -> Option<String> {
-    find_linux_desktop_icon_name(entry, detected_path)
-        .and_then(|name| linux_icon_file_from_name(&name))
-        .and_then(|path| image_data_url_from_file(&path))
-}
-
-/// Walk up from `detected_path` looking for any of the manifest's
-/// `windows_exe_names`. Used to recover the real `.exe` (which carries
-/// the embedded icon resource) from PATH-resolved shims like
-/// `bin/code.cmd` (one level up) or `resources/app/bin/cursor.cmd`
-/// (three levels up). Falls back to `detected_path` so apps without
-/// a manifest entry still get whatever icon ExtractAssociatedIcon can
-/// produce.
-#[cfg(target_os = "windows")]
-fn resolve_windows_icon_source(entry: &AppEntry, detected_path: &Path) -> PathBuf {
-    if entry.windows_exe_names.is_empty() {
-        return detected_path.to_path_buf();
-    }
-    // Five levels covers every layout we ship in default-apps.json
-    // (Cursor's three-deep shim is the deepest case) without risking
-    // a runaway directory walk on weird filesystems.
-    let mut dir = detected_path.parent();
-    for _ in 0..5 {
-        let Some(d) = dir else { break };
-        for exe in &entry.windows_exe_names {
-            let candidate = d.join(exe);
-            if candidate.is_file() {
-                return candidate;
-            }
-        }
-        dir = d.parent();
-    }
-    detected_path.to_path_buf()
-}
-
-/// Inline C# helper compiled at PowerShell startup. Three extraction
-/// paths in priority order:
-///
-/// 1. **AppxPackage logo** (`windows_appx_package` set) — UWP/MSIX
-///    apps whose real binary is in `%PROGRAMFILES%\WindowsApps` and
-///    is reached through a 0-byte execution alias on PATH. The alias
-///    short-circuits `IShellItemImageFactory` to the console
-///    subsystem (Windows Terminal's `wt.exe` ends up indistinguishable
-///    from `cmd.exe`). `Get-AppxPackage` exposes the package's
-///    `InstallLocation`; the manifest's `Assets/*Logo*.png` files are
-///    the real vendor logos at full resolution. We pick the largest
-///    PNG by file size and skip variants that don't render in dark
-///    UIs (`*altform-unplated*`, `*contrast-*`).
-/// 2. **`IShellItemImageFactory::GetImage` at 256×256** — the same
-///    shell API Explorer uses for its "Extra large icons" view.
-///    `SIIGBF_ICONONLY` keeps us from accidentally getting a PE-header
-///    thumbnail; `SIIGBF_BIGGERSIZEOK` lets the shell hand back its
-///    largest cached variant if 256 isn't directly available.
-/// 3. **`ExtractAssociatedIcon`** — last-resort 32×32 grab. Some
-///    shimmed paths fail the COM call entirely; this keeps the menu
-///    from showing the lucide placeholder for them.
-///
-/// Input is read from stdin: line 1 is the AppxPackage name (empty
-/// string skips the UWP path), line 2 is the .exe path. Stdin avoids
-/// every PowerShell `-Command` quoting hazard.
-#[cfg(target_os = "windows")]
-const ICON_EXTRACT_PS1: &str = r#"
-$ErrorActionPreference = 'Stop'
-$src = @'
-using System;
-using System.Drawing;
-using System.Drawing.Imaging;
-using System.IO;
-using System.Runtime.InteropServices;
-public class ClaudetteIcon {
-    [StructLayout(LayoutKind.Sequential)]
-    public struct SIZE { public int cx, cy; }
-
-    [ComImport, Guid("BCC18B79-BA16-442F-80C4-8A59C30C463B"),
-     InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    public interface IShellItemImageFactory {
-        [PreserveSig] int GetImage(SIZE size, int flags, out IntPtr phbm);
-    }
-
-    [DllImport("shell32.dll", CharSet=CharSet.Unicode, SetLastError=true)]
-    static extern int SHCreateItemFromParsingName(
-        [MarshalAs(UnmanagedType.LPWStr)] string path,
-        IntPtr bc,
-        ref Guid riid,
-        [Out, MarshalAs(UnmanagedType.Interface)] out IShellItemImageFactory ppv);
-
-    [DllImport("gdi32.dll")] static extern bool DeleteObject(IntPtr h);
-
-    public static byte[] ShellImage(string path, int size) {
-        Guid iid = new Guid("BCC18B79-BA16-442F-80C4-8A59C30C463B");
-        IShellItemImageFactory factory;
-        int hr = SHCreateItemFromParsingName(path, IntPtr.Zero, ref iid, out factory);
-        if (hr != 0) return null;
-        IntPtr hbm;
-        SIZE sz; sz.cx = size; sz.cy = size;
-        hr = factory.GetImage(sz, 0x4 | 0x1, out hbm);
-        if (hr != 0) return null;
-        try {
-            using (Bitmap bmp = Bitmap.FromHbitmap(hbm))
-            using (Bitmap argb = new Bitmap(bmp.Width, bmp.Height, PixelFormat.Format32bppArgb)) {
-                using (Graphics g = Graphics.FromImage(argb)) g.DrawImage(bmp, 0, 0);
-                using (MemoryStream ms = new MemoryStream()) {
-                    argb.Save(ms, ImageFormat.Png);
-                    return ms.ToArray();
-                }
-            }
-        } finally { DeleteObject(hbm); }
-    }
-
-    public static byte[] AssociatedIcon(string path) {
-        using (Icon ico = Icon.ExtractAssociatedIcon(path)) {
-            if (ico == null) return null;
-            using (Bitmap bmp = ico.ToBitmap())
-            using (MemoryStream ms = new MemoryStream()) {
-                bmp.Save(ms, ImageFormat.Png);
-                return ms.ToArray();
-            }
-        }
-    }
-}
-'@
-Add-Type -TypeDefinition $src -ReferencedAssemblies System.Drawing | Out-Null
-$pkg  = [Console]::In.ReadLine()
-$path = [Console]::In.ReadLine()
-
-# UWP path: read the AppxManifest to learn the *declared* logo
-# (any of the Square*Logo / Logo entries), then glob the same
-# directory for size-variant siblings (`Square44x44Logo.targetsize-256.png`,
-# `*.scale-400.png`, …) and keep the largest. The declared filename
-# rarely exists on disk by itself — Windows splits each logo into
-# one PNG per scale + target size at install time, leaving only the
-# variants. We filter out `altform-unplated` (transparent badges
-# that look wrong on a dark menu background) and `contrast-` (high-
-# contrast accessibility variants). File size is the biggest-icon
-# heuristic — UWP logos are anti-aliased PNGs whose byte count
-# scales with pixel count.
-#
-# `Get-AppxPackageManifest` works for unprivileged users, and
-# `Get-ChildItem` *can* enumerate per-package install dirs under
-# `%PROGRAMFILES%\WindowsApps` despite the parent being locked
-# down — each package's ACL grants the user list rights to its own
-# subtree.
-if (-not [string]::IsNullOrEmpty($pkg)) {
-    $appx = Get-AppxPackage -Name "${pkg}*" -ErrorAction SilentlyContinue | Select-Object -First 1
-    if ($appx) {
-        $manifest = $null
-        try { $manifest = Get-AppxPackageManifest -Package $appx.PackageFullName -ErrorAction Stop } catch { }
-        $logoRels = @()
-        if ($manifest) {
-            $vis = $manifest.Package.Applications.Application.VisualElements
-            if ($vis) {
-                foreach ($attr in 'Square150x150Logo','Square44x44Logo','Square71x71Logo','Logo') {
-                    $val = $vis.$attr
-                    if ($val) { $logoRels += $val }
-                }
-            }
-            $propLogo = $manifest.Package.Properties.Logo
-            if ($propLogo) { $logoRels += $propLogo }
-        }
-        # Last-resort guesses if the manifest didn't yield anything
-        # parseable. Most UWP packages put logos under one of these.
-        if ($logoRels.Count -eq 0) { $logoRels = @('Images\Logo.png','Assets\Logo.png') }
-        $best = $null
-        foreach ($rel in $logoRels) {
-            $fullRel = Join-Path $appx.InstallLocation $rel
-            $stem = [IO.Path]::GetFileNameWithoutExtension($fullRel)
-            $dir  = [IO.Path]::GetDirectoryName($fullRel)
-            if (-not (Test-Path $dir)) { continue }
-            $candidate = Get-ChildItem -Path $dir -Filter "$stem*.png" -ErrorAction SilentlyContinue |
-                Where-Object { $_.Name -notmatch '(?i)altform-unplated|contrast-' } |
-                Sort-Object Length -Descending |
-                Select-Object -First 1
-            if ($candidate -and (-not $best -or $candidate.Length -gt $best.Length)) {
-                $best = $candidate
-            }
-        }
-        if ($best) {
-            [Convert]::ToBase64String([IO.File]::ReadAllBytes($best.FullName))
-            exit 0
-        }
-    }
-}
-
-if ([string]::IsNullOrEmpty($path)) { exit 2 }
-$bytes = [ClaudetteIcon]::ShellImage($path, 256)
-if ($null -eq $bytes -or $bytes.Length -eq 0) {
-    $bytes = [ClaudetteIcon]::AssociatedIcon($path)
-}
-if ($null -eq $bytes -or $bytes.Length -eq 0) { exit 3 }
-[Convert]::ToBase64String($bytes)
-"#;
-
-/// Where the icon disk cache lives. Routed through
-/// [`claudette::path::claudette_home`] so it follows the same
-/// `$CLAUDETTE_HOME` override that `apps_config_path` and the rest
-/// of the on-disk state honor — important for `dev --new` and
-/// `dev --clone`, which point `CLAUDETTE_HOME` at a per-PID tmp
-/// sandbox: cache entries land under the sandbox and get cleaned
-/// up when it tears down, instead of leaking into the real
-/// `~/.claudette/icon-cache/` and surviving across sessions.
-///
-/// Returns `Option` purely for symmetry with the old shape (callers
-/// already handle `None`); `claudette_home` itself falls back to
-/// `./.claudette` rather than failing, so this is effectively
-/// infallible today.
-#[cfg(target_os = "windows")]
-fn icon_cache_dir() -> Option<PathBuf> {
-    Some(claudette::path::claudette_home().join("icon-cache"))
-}
-
-/// Build a stable cache key for a Windows icon source.
-///
-/// For .exe paths we mix the canonical path, file size, and modification
-/// time so an in-place upgrade (e.g. VS Code auto-updating `Code.exe`)
-/// invalidates the cache automatically — the cache filename changes,
-/// the new run misses, and the new icon is extracted and stored under
-/// a new key.
-///
-/// For UWP packages the .exe is the alias so the path is identical
-/// across versions; we use just the package-name field. AppxPackage
-/// upgrades therefore won't auto-refresh — but Windows Terminal's
-/// logo hasn't changed in years and the cost of a stale icon is
-/// purely cosmetic. If a user wants a refresh they can clear
-/// `~/.claudette/icon-cache/`.
-///
-/// `windows_appx_package` takes precedence to ensure we cache the
-/// UWP-resolved icon (which the .exe path can't produce) under a
-/// key that doesn't collide with the .exe-derived key.
-#[cfg(target_os = "windows")]
-fn icon_cache_key(appx_package: &str, icon_source: &Path) -> Option<String> {
-    use std::collections::hash_map::DefaultHasher;
-    use std::hash::{Hash, Hasher};
-
-    let mut hasher = DefaultHasher::new();
-    if !appx_package.is_empty() {
-        "appx".hash(&mut hasher);
-        appx_package.hash(&mut hasher);
-    } else {
-        "exe".hash(&mut hasher);
-        icon_source.hash(&mut hasher);
-        if let Ok(meta) = std::fs::metadata(icon_source) {
-            meta.len().hash(&mut hasher);
-            if let Ok(mtime) = meta.modified()
-                && let Ok(dur) = mtime.duration_since(std::time::UNIX_EPOCH)
-            {
-                dur.as_secs().hash(&mut hasher);
-            }
-        } else {
-            // Missing source — refuse to cache so the next attempt
-            // re-runs extraction once the file appears.
-            return None;
-        }
-    }
-    Some(format!("{:016x}", hasher.finish()))
-}
-
-#[cfg(target_os = "windows")]
-fn read_icon_cache(key: &str) -> Option<String> {
-    let path = icon_cache_dir()?.join(format!("{key}.b64"));
-    let data = std::fs::read_to_string(&path).ok()?;
-    let trimmed = data.trim();
-    (!trimmed.is_empty() && trimmed.starts_with("data:image/")).then(|| trimmed.to_owned())
-}
-
-#[cfg(target_os = "windows")]
-fn write_icon_cache(key: &str, data_url: &str) {
-    let Some(dir) = icon_cache_dir() else { return };
-    let _ = std::fs::create_dir_all(&dir);
-    let _ = std::fs::write(dir.join(format!("{key}.b64")), data_url);
-}
-
-#[cfg(target_os = "windows")]
-fn app_icon_data_url(entry: &AppEntry, detected_path: &Path) -> Option<String> {
-    let icon_source = resolve_windows_icon_source(entry, detected_path);
-
-    // Cache hit: the only case where extraction is *truly* free —
-    // no PowerShell startup, no Add-Type compile, no COM call. This
-    // is what makes "open the dropdown" feel instant on every launch
-    // after the first.
-    let cache_key = icon_cache_key(&entry.windows_appx_package, &icon_source);
-    if let Some(key) = cache_key.as_deref()
-        && let Some(cached) = read_icon_cache(key)
-    {
-        return Some(cached);
-    }
-
-    let data_url = extract_windows_icon_data_url(&entry.windows_appx_package, &icon_source)?;
-
-    if let Some(key) = cache_key.as_deref() {
-        write_icon_cache(key, &data_url);
-    }
-    Some(data_url)
-}
-
-/// Spawn a `powershell.exe` child, feed the AppxPackage name + target
-/// path on stdin (two newline-separated lines), and parse the base64
-/// PNG it prints. Lifted out of `app_icon_data_url` so the regression
-/// tests can exercise the extractor directly without manufacturing an
-/// `AppEntry` + invoking the cache.
-#[cfg(target_os = "windows")]
-fn extract_windows_icon_data_url(appx_package: &str, icon_source: &Path) -> Option<String> {
-    use std::io::Write;
-    use std::process::Stdio;
-
-    let mut child = claudette::process::std_command("powershell.exe")
-        .args([
-            "-NoProfile",
-            "-NonInteractive",
-            "-ExecutionPolicy",
-            "Bypass",
-            "-Command",
-            ICON_EXTRACT_PS1,
-        ])
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::null())
-        .spawn()
-        .ok()?;
-
-    {
-        let stdin = child.stdin.as_mut()?;
-        // Two lines: AppxPackage name (or empty) then path. Empty
-        // lines are deliberate — PowerShell's
-        // `[Console]::In.ReadLine()` returns `""` for them, which
-        // the script normalizes via `IsNullOrEmpty`.
-        let _ = writeln!(stdin, "{appx_package}");
-        let _ = writeln!(stdin, "{}", icon_source.display());
-    }
-    let output = child.wait_with_output().ok()?;
-    if !output.status.success() {
-        return None;
-    }
-
-    let encoded = String::from_utf8(output.stdout).ok()?;
-    let encoded = encoded.trim();
-    (!encoded.is_empty()).then(|| format!("data:image/png;base64,{encoded}"))
-}
-
-#[cfg(not(any(target_os = "macos", target_os = "linux", target_os = "windows")))]
-fn app_icon_data_url(_entry: &AppEntry, _detected_path: &Path) -> Option<String> {
-    None
-}
-
-/// Detect installed apps from the given config, searching the provided PATH dirs.
-/// This is the testable core — `detect_from_config` wraps it with the real PATH.
-fn detect_with_paths(config: &AppsConfig, path_dirs: &[PathBuf]) -> Vec<DetectedApp> {
-    let category_order = |c: &AppCategory| -> u8 {
-        match c {
-            AppCategory::Editor => 0,
-            AppCategory::FileManager => 1,
-            AppCategory::Terminal => 2,
-            AppCategory::Ide => 3,
-        }
-    };
-
-    let mut detected: Vec<(usize, DetectedApp)> = Vec::new();
-
-    for (index, entry) in config.apps.iter().enumerate() {
-        // Try bin_names first.
-        if let Some(bin_path) = entry
-            .bin_names
-            .iter()
-            .find_map(|name| find_binary(name, path_dirs))
-        {
-            let icon_data_url = app_icon_data_url(entry, &bin_path);
-            detected.push((
-                index,
-                DetectedApp {
-                    id: entry.id.clone(),
-                    name: entry.name.clone(),
-                    category: entry.category,
-                    detected_path: bin_path.to_string_lossy().to_string(),
-                    icon_data_url,
-                },
-            ));
-            continue;
-        }
-
-        // Try mac_app_names (macOS only).
-        #[cfg(target_os = "macos")]
-        if let Some(app_path) = entry
-            .mac_app_names
-            .iter()
-            .find_map(|name| find_mac_app(name))
-        {
-            let icon_data_url = app_icon_data_url(entry, &app_path);
-            detected.push((
-                index,
-                DetectedApp {
-                    id: entry.id.clone(),
-                    name: entry.name.clone(),
-                    category: entry.category,
-                    detected_path: app_path.to_string_lossy().to_string(),
-                    icon_data_url,
-                },
-            ));
-            continue;
-        }
-    }
-
-    detected.sort_by(|a, b| {
-        category_order(&a.1.category)
-            .cmp(&category_order(&b.1.category))
-            .then_with(|| a.0.cmp(&b.0))
-    });
-
-    detected.into_iter().map(|(_, app)| app).collect()
-}
-
-/// Public detection entry point using the real system PATH.
-fn detect_from_config(config: &AppsConfig) -> Vec<DetectedApp> {
-    let path_dirs = build_path_dirs();
-    detect_with_paths(config, &path_dirs)
-}
-
-// ---------------------------------------------------------------------------
-// Tauri commands
-// ---------------------------------------------------------------------------
+use config::load_apps_config;
+use detection::detect_from_config;
+pub(crate) use launch::open_workspace_in_app_inner;
+#[allow(unused_imports)]
+pub use model::{AppCategory, AppEntry, AppsConfig, DetectedApp};
+pub(crate) use model::{DEFAULT_TERMINAL_APP_SETTING_KEY, select_workspace_terminal_app_id};
 
 #[tauri::command]
 pub async fn detect_installed_apps(state: State<'_, AppState>) -> Result<Vec<DetectedApp>, String> {
@@ -1091,316 +32,6 @@ pub async fn detect_installed_apps(state: State<'_, AppState>) -> Result<Vec<Det
     Ok(apps)
 }
 
-/// Launch an app using macOS `open -a` command.
-#[cfg(target_os = "macos")]
-async fn open_macos_app(app_name: &str, target_path: &str) -> Result<(), String> {
-    claudette::process::command("open")
-        .args(["-a", app_name, target_path])
-        .spawn()
-        .map_err(|e| format!("Failed to launch {app_name}: {e}"))?;
-    Ok(())
-}
-
-/// Launch a terminal app via AppleScript (iTerm2, Terminal.app).
-/// Uses `on run argv` + `quoted form of` to pass the path as an argument,
-/// avoiding string interpolation and AppleScript injection risks.
-#[cfg(target_os = "macos")]
-async fn open_applescript(app_id: &str, worktree_path: &str) -> Result<(), String> {
-    let script = match app_id {
-        "iterm2" => {
-            r#"on run argv
-    set p to item 1 of argv
-    set cmd to "cd " & quoted form of p & " && exec $SHELL"
-    tell application "iTerm"
-        activate
-        if (count of windows) = 0 then
-            create window with default profile command cmd
-        else
-            tell current window
-                set newTab to (create tab with default profile)
-                tell current session of newTab
-                    write text cmd
-                end tell
-            end tell
-        end if
-    end tell
-end run"#
-        }
-        "macos-terminal" => {
-            r#"on run argv
-    set p to item 1 of argv
-    set cmd to "cd " & quoted form of p
-    tell application "Terminal"
-        activate
-        do script cmd
-    end tell
-end run"#
-        }
-        other => return Err(format!("No AppleScript handler for app '{other}'")),
-    };
-
-    claudette::process::command("osascript")
-        .arg("-e")
-        .arg(script)
-        .arg("--")
-        .arg(worktree_path)
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .spawn()
-        .map_err(|e| format!("Failed to run AppleScript for {app_id}: {e}"))?;
-    Ok(())
-}
-
-/// Determine the exec-separator args for launching an editor inside a given terminal.
-fn terminal_exec_args(terminal_id: &str) -> &'static [&'static str] {
-    match terminal_id {
-        "alacritty" | "konsole" | "xfce4-terminal" => &["-e"],
-        "gnome-terminal" => &["--"],
-        // kitty, foot, wezterm, ghostty: just append the command directly.
-        _ => &[],
-    }
-}
-
-/// Shell-quote a string using single quotes (POSIX-safe).
-/// e.g. `hello world` → `'hello world'`, `it's` → `'it'\''s'`
-#[cfg(target_os = "macos")]
-fn shell_quote(s: &str) -> String {
-    format!("'{}'", s.replace('\'', "'\\''"))
-}
-
-/// Launch a TUI editor via AppleScript when only .app-bundle terminals are available (macOS).
-/// Builds a fully shell-quoted command in Rust and passes it as a single argument
-/// to avoid injection risks from paths or args with special characters.
-#[cfg(target_os = "macos")]
-async fn open_tui_via_applescript(
-    editor_entry: &AppEntry,
-    editor_detected: &DetectedApp,
-    worktree_path: &str,
-    terminal: &DetectedApp,
-) -> Result<(), String> {
-    // Build a properly shell-quoted command: cd '<path>' && '<editor>' '<arg1>' '<arg2>' ...
-    let mut editor_parts = vec![shell_quote(&editor_detected.detected_path)];
-    for arg in &editor_entry.open_args {
-        editor_parts.push(shell_quote(&arg.replace("{}", ".")));
-    }
-    let full_cmd = format!(
-        "cd {} && {}",
-        shell_quote(worktree_path),
-        editor_parts.join(" ")
-    );
-
-    let (app_name, script) = if terminal.id == "iterm2" {
-        (
-            "iTerm",
-            r#"on run argv
-    set cmd to item 1 of argv
-    tell application "iTerm"
-        activate
-        if (count of windows) = 0 then
-            create window with default profile command cmd
-        else
-            tell current window
-                set newTab to (create tab with default profile)
-                tell current session of newTab
-                    write text cmd
-                end tell
-            end tell
-        end if
-    end tell
-end run"#,
-        )
-    } else {
-        (
-            "Terminal",
-            r#"on run argv
-    set cmd to item 1 of argv
-    tell application "Terminal"
-        activate
-        do script cmd
-    end tell
-end run"#,
-        )
-    };
-
-    claudette::process::command("osascript")
-        .arg("-e")
-        .arg(script)
-        .arg("--")
-        .arg(&full_cmd)
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .spawn()
-        .map_err(|e| {
-            format!(
-                "Failed to launch {} in {app_name} via AppleScript: {e}",
-                editor_entry.name
-            )
-        })?;
-    Ok(())
-}
-
-/// Launch a TUI editor (needs_terminal=true) inside the best detected terminal.
-/// Prefers terminals detected via binary path; falls back to AppleScript on macOS
-/// when only .app-bundle terminals are available.
-async fn open_in_terminal(
-    editor_entry: &AppEntry,
-    editor_detected: &DetectedApp,
-    worktree_path: &str,
-    state: &AppState,
-) -> Result<(), String> {
-    let config = load_apps_config();
-
-    let detected_apps = state.detected_apps.read().await;
-
-    // Prefer a terminal detected via a real binary (not a .app bundle),
-    // because .app paths can't be passed to Command::new directly.
-    let terminal = detected_apps
-        .iter()
-        .filter(|a| a.category == AppCategory::Terminal)
-        .find(|a| !a.detected_path.ends_with(".app"))
-        .or_else(|| {
-            detected_apps
-                .iter()
-                .find(|a| a.category == AppCategory::Terminal)
-        })
-        .ok_or("No terminal emulator detected — cannot launch TUI editor")?
-        .clone();
-    drop(detected_apps);
-
-    // If the terminal was detected via .app bundle, use AppleScript on macOS.
-    #[cfg(target_os = "macos")]
-    if terminal.detected_path.ends_with(".app") {
-        return open_tui_via_applescript(editor_entry, editor_detected, worktree_path, &terminal)
-            .await;
-    }
-
-    let terminal_entry = config
-        .apps
-        .iter()
-        .find(|a| a.id == terminal.id)
-        .ok_or_else(|| format!("Terminal '{}' not found in config", terminal.id))?;
-
-    // Build: terminal_binary [terminal_open_args with {} -> path] [exec_separator] editor_binary [editor_open_args]
-    let mut cmd = claudette::process::command(&terminal.detected_path);
-    #[cfg(target_os = "windows")]
-    cmd.new_console_window();
-
-    for arg in &terminal_entry.open_args {
-        cmd.arg(arg.replace("{}", worktree_path));
-    }
-
-    for arg in terminal_exec_args(&terminal.id) {
-        cmd.arg(arg);
-    }
-
-    cmd.arg(&editor_detected.detected_path);
-    for arg in &editor_entry.open_args {
-        cmd.arg(arg.replace("{}", "."));
-    }
-
-    cmd.spawn().map_err(|e| {
-        format!(
-            "Failed to launch {} in {}: {e}",
-            editor_entry.name, terminal.name
-        )
-    })?;
-    Ok(())
-}
-
-pub(crate) async fn open_workspace_in_app_inner(
-    app_id: &str,
-    worktree_path: &str,
-    state: &AppState,
-) -> Result<(), String> {
-    // Reload config each time so edits to open_args, needs_terminal, etc. take
-    // effect without restart.  Note: the *detected apps list* (which apps appear
-    // in the menu) is cached from startup; adding a new app requires restart.
-    let config = load_apps_config();
-    let entry = config
-        .apps
-        .iter()
-        .find(|a| a.id == app_id)
-        .ok_or_else(|| format!("App '{app_id}' not found in apps.json"))?
-        .clone();
-
-    // Handle AppleScript sentinel (iTerm2, Terminal.app).
-    #[cfg(target_os = "macos")]
-    if entry
-        .open_args
-        .first()
-        .is_some_and(|a| a == "__applescript__")
-    {
-        return open_applescript(app_id, worktree_path).await;
-    }
-
-    #[cfg(target_os = "macos")]
-    if entry.open_args.first().is_some_and(|a| a == "__open__") {
-        claudette::process::command("open")
-            .arg(worktree_path)
-            .spawn()
-            .map_err(|e| format!("Failed to open workspace: {e}"))?;
-        return Ok(());
-    }
-
-    // Handle __open_a__ sentinel (Xcode) — look up detected_path to get the .app bundle.
-    #[cfg(target_os = "macos")]
-    if entry.open_args.first().is_some_and(|a| a == "__open_a__") {
-        let detected_apps = state.detected_apps.read().await;
-        let detected = detected_apps
-            .iter()
-            .find(|a| a.id == app_id)
-            .ok_or_else(|| format!("App '{app_id}' not detected on this system"))?;
-        let app_path = detected.detected_path.clone();
-        drop(detected_apps);
-        return open_macos_app(&app_path, worktree_path).await;
-    }
-
-    // Look up the detected path for this app.
-    let detected_apps = state.detected_apps.read().await;
-    let detected = detected_apps
-        .iter()
-        .find(|a| a.id == app_id)
-        .ok_or_else(|| format!("App '{app_id}' not detected on this system"))?
-        .clone();
-    drop(detected_apps);
-
-    // Handle TUI editors that need a terminal host.
-    if entry.needs_terminal {
-        return open_in_terminal(&entry, &detected, worktree_path, state).await;
-    }
-
-    // Handle .app-only detection on macOS (CLI not in PATH).
-    #[cfg(target_os = "macos")]
-    if detected.detected_path.ends_with(".app") {
-        return open_macos_app(&detected.detected_path, worktree_path).await;
-    }
-
-    // Normal binary launch: substitute {} in open_args with the worktree path.
-    let args: Vec<String> = entry
-        .open_args
-        .iter()
-        .map(|a| a.replace("{}", worktree_path))
-        .collect();
-
-    let mut cmd = claudette::process::command(&detected.detected_path);
-    // Windows console terminals (cmd.exe, powershell.exe, pwsh.exe)
-    // need a fresh, visible console of their own. The process helper
-    // suppresses console allocation by default for everything else.
-    // `wt.exe` ignores this flag because it activates the Windows Terminal
-    // app via a separate process, so applying it to the Terminal category
-    // is harmless there.
-    #[cfg(target_os = "windows")]
-    if entry.category == AppCategory::Terminal {
-        cmd.new_console_window();
-    }
-
-    cmd.args(&args)
-        .spawn()
-        .map_err(|e| format!("Failed to launch {}: {e}", entry.name))?;
-
-    Ok(())
-}
-
 #[tauri::command]
 pub async fn open_workspace_in_app(
     app_id: String,
@@ -1408,1138 +39,4 @@ pub async fn open_workspace_in_app(
     state: State<'_, AppState>,
 ) -> Result<(), String> {
     open_workspace_in_app_inner(&app_id, &worktree_path, state.inner()).await
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn detected_app(id: &str, name: &str, category: AppCategory) -> DetectedApp {
-        DetectedApp {
-            id: id.to_string(),
-            name: name.to_string(),
-            category,
-            detected_path: format!("/usr/bin/{id}"),
-            icon_data_url: None,
-        }
-    }
-
-    #[test]
-    fn parse_valid_config() {
-        let json = r#"{
-            "apps": [{
-                "id": "test-editor",
-                "name": "Test Editor",
-                "category": "editor",
-                "bin_names": ["testedit"],
-                "mac_app_names": ["Test Editor.app"],
-                "open_args": ["{}"],
-                "needs_terminal": false
-            }]
-        }"#;
-        let config: AppsConfig = serde_json::from_str(json).unwrap();
-        assert_eq!(config.apps.len(), 1);
-        assert_eq!(config.apps[0].id, "test-editor");
-        assert_eq!(config.apps[0].name, "Test Editor");
-        assert_eq!(config.apps[0].category, AppCategory::Editor);
-        assert_eq!(config.apps[0].bin_names, vec!["testedit"]);
-        assert_eq!(config.apps[0].open_args, vec!["{}"]);
-        assert!(!config.apps[0].needs_terminal);
-    }
-
-    #[test]
-    fn parse_optional_fields_use_defaults() {
-        let json = r#"{
-            "apps": [{
-                "id": "minimal",
-                "name": "Minimal",
-                "category": "terminal",
-                "open_args": ["--dir", "{}"]
-            }]
-        }"#;
-        let config: AppsConfig = serde_json::from_str(json).unwrap();
-        let app = &config.apps[0];
-        assert!(app.bin_names.is_empty());
-        assert!(app.mac_app_names.is_empty());
-        assert!(!app.needs_terminal);
-    }
-
-    #[test]
-    fn parse_malformed_json_is_err() {
-        let result = serde_json::from_str::<AppsConfig>("not valid json {{{");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_unknown_fields_ignored() {
-        let json = r#"{
-            "apps": [{
-                "id": "x",
-                "name": "X",
-                "category": "ide",
-                "open_args": ["{}"],
-                "future_field": true,
-                "another": 42
-            }],
-            "version": 99
-        }"#;
-        let config: AppsConfig = serde_json::from_str(json).unwrap();
-        assert_eq!(config.apps[0].id, "x");
-        assert_eq!(config.apps[0].category, AppCategory::Ide);
-    }
-
-    #[test]
-    fn parse_file_manager_category() {
-        let json = r#"{
-            "apps": [{
-                "id": "finder",
-                "name": "Finder",
-                "category": "file_manager",
-                "open_args": ["__open__"]
-            }]
-        }"#;
-        let config: AppsConfig = serde_json::from_str(json).unwrap();
-        assert_eq!(config.apps[0].category, AppCategory::FileManager);
-    }
-
-    #[test]
-    fn select_workspace_terminal_prefers_detected_override() {
-        let apps = vec![
-            detected_app("vscode", "VS Code", AppCategory::Editor),
-            detected_app("terminal", "Terminal", AppCategory::Terminal),
-            detected_app("ghostty", "Ghostty", AppCategory::Terminal),
-        ];
-
-        assert_eq!(
-            select_workspace_terminal_app_id(&apps, Some("ghostty")).as_deref(),
-            Some("ghostty")
-        );
-    }
-
-    #[test]
-    fn select_workspace_terminal_ignores_non_terminal_override() {
-        let apps = vec![
-            detected_app("vscode", "VS Code", AppCategory::Editor),
-            detected_app("terminal", "Terminal", AppCategory::Terminal),
-        ];
-
-        assert_eq!(
-            select_workspace_terminal_app_id(&apps, Some("vscode")).as_deref(),
-            Some("terminal")
-        );
-    }
-
-    #[test]
-    fn select_workspace_terminal_uses_first_detected_for_auto() {
-        let apps = vec![
-            detected_app("vscode", "VS Code", AppCategory::Editor),
-            detected_app("iterm2", "iTerm2", AppCategory::Terminal),
-            detected_app("ghostty", "Ghostty", AppCategory::Terminal),
-        ];
-
-        assert_eq!(
-            select_workspace_terminal_app_id(&apps, None).as_deref(),
-            Some("iterm2")
-        );
-    }
-
-    #[test]
-    fn select_workspace_terminal_returns_none_without_terminals() {
-        let apps = vec![detected_app("vscode", "VS Code", AppCategory::Editor)];
-
-        assert_eq!(
-            select_workspace_terminal_app_id(&apps, Some("ghostty")),
-            None
-        );
-    }
-
-    #[test]
-    fn parse_embedded_default_config() {
-        let config: AppsConfig =
-            serde_json::from_str(DEFAULT_APPS_JSON).expect("default-apps.json must parse");
-        assert!(config.apps.len() >= 15, "expected at least 15 default apps");
-        // Spot-check a few entries
-        assert!(config.apps.iter().any(|a| a.id == "vscode"));
-        assert!(config.apps.iter().any(|a| a.id == "ghostty"));
-        assert!(
-            config
-                .apps
-                .iter()
-                .any(|a| a.id == "neovim" && a.needs_terminal)
-        );
-    }
-
-    #[test]
-    fn load_apps_config_missing_file_returns_default() {
-        let tmp = tempfile::tempdir().unwrap();
-        let path = tmp.path().join("apps.json");
-        let config = load_apps_config_from(&path);
-        assert!(!config.apps.is_empty());
-    }
-
-    #[test]
-    fn load_apps_config_malformed_file_returns_default() {
-        let tmp = tempfile::tempdir().unwrap();
-        let path = tmp.path().join("apps.json");
-        std::fs::write(&path, "NOT JSON").unwrap();
-        let config = load_apps_config_from(&path);
-        assert!(!config.apps.is_empty());
-    }
-
-    #[test]
-    fn detect_finds_executable_in_path() {
-        let tmp = tempfile::tempdir().unwrap();
-        let bin = tmp.path().join("myeditor");
-        std::fs::write(&bin, "#!/bin/sh\n").unwrap();
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o755)).unwrap();
-        }
-
-        let config = AppsConfig {
-            apps: vec![AppEntry {
-                id: "myeditor".into(),
-                name: "My Editor".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["myeditor".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec![],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-
-        let detected = detect_with_paths(&config, &[tmp.path().to_path_buf()]);
-        assert_eq!(detected.len(), 1);
-        assert_eq!(detected[0].id, "myeditor");
-        assert_eq!(detected[0].name, "My Editor");
-        assert_eq!(detected[0].category, AppCategory::Editor);
-        assert_eq!(detected[0].detected_path, bin.to_string_lossy().to_string());
-    }
-
-    #[test]
-    fn detect_skips_missing_binary() {
-        let tmp = tempfile::tempdir().unwrap();
-        // No binary created — the directory is empty.
-        let config = AppsConfig {
-            apps: vec![AppEntry {
-                id: "missing".into(),
-                name: "Missing App".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["nonexistent-binary".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec![],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-
-        let detected = detect_with_paths(&config, &[tmp.path().to_path_buf()]);
-        assert!(detected.is_empty());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn detect_skips_non_executable_file() {
-        let tmp = tempfile::tempdir().unwrap();
-        let bin = tmp.path().join("noexec");
-        std::fs::write(&bin, "data").unwrap();
-        // Permissions 0o644 — not executable.
-        use std::os::unix::fs::PermissionsExt;
-        std::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o644)).unwrap();
-
-        let config = AppsConfig {
-            apps: vec![AppEntry {
-                id: "noexec".into(),
-                name: "No Exec".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["noexec".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec![],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-
-        let detected = detect_with_paths(&config, &[tmp.path().to_path_buf()]);
-        assert!(detected.is_empty());
-    }
-
-    #[test]
-    fn detect_sorted_by_category_then_config_order() {
-        let tmp = tempfile::tempdir().unwrap();
-        for name in ["zterm", "beditor", "aeditor"] {
-            let bin = tmp.path().join(name);
-            std::fs::write(&bin, "#!/bin/sh\n").unwrap();
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                std::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o755)).unwrap();
-            }
-        }
-
-        let config = AppsConfig {
-            apps: vec![
-                AppEntry {
-                    id: "zterm".into(),
-                    name: "Z Terminal".into(),
-                    category: AppCategory::Terminal,
-                    bin_names: vec!["zterm".into()],
-                    mac_app_names: vec![],
-                    windows_exe_names: vec![],
-                    windows_appx_package: String::new(),
-                    open_args: vec!["{}".into()],
-                    needs_terminal: false,
-                },
-                AppEntry {
-                    id: "beditor".into(),
-                    name: "B Editor".into(),
-                    category: AppCategory::Editor,
-                    bin_names: vec!["beditor".into()],
-                    mac_app_names: vec![],
-                    windows_exe_names: vec![],
-                    windows_appx_package: String::new(),
-                    open_args: vec!["{}".into()],
-                    needs_terminal: false,
-                },
-                AppEntry {
-                    id: "aeditor".into(),
-                    name: "A Editor".into(),
-                    category: AppCategory::Editor,
-                    bin_names: vec!["aeditor".into()],
-                    mac_app_names: vec![],
-                    windows_exe_names: vec![],
-                    windows_appx_package: String::new(),
-                    open_args: vec!["{}".into()],
-                    needs_terminal: false,
-                },
-            ],
-        };
-
-        let detected = detect_with_paths(&config, &[tmp.path().to_path_buf()]);
-        assert_eq!(detected.len(), 3);
-        assert_eq!(
-            detected
-                .iter()
-                .map(|app| app.id.as_str())
-                .collect::<Vec<_>>(),
-            vec!["beditor", "aeditor", "zterm"],
-        );
-    }
-
-    /// Loading a stale user `apps.json` should yield a config that
-    /// includes any embedded entries the user file is missing. This is
-    /// the migration path for users whose file pre-dates the addition
-    /// of e.g. Windows Terminal, PowerShell 7, etc.
-    #[test]
-    fn load_apps_config_merges_missing_default_entries() {
-        let tmp = tempfile::tempdir().unwrap();
-        let path = tmp.path().join("apps.json");
-        // Write a one-entry user config — nothing else present.
-        std::fs::write(
-            &path,
-            r#"{"apps":[{"id":"vscode","name":"VS Code","category":"editor","bin_names":["code"],"open_args":["{}"]}]}"#,
-        )
-        .unwrap();
-
-        let config = load_apps_config_from(&path);
-
-        // User's vscode entry is preserved.
-        assert!(config.apps.iter().any(|a| a.id == "vscode"));
-        // Embedded entries the user lacked are now present.
-        assert!(config.apps.iter().any(|a| a.id == "cursor"));
-        assert!(config.apps.iter().any(|a| a.id == "windows-terminal"));
-        assert!(config.apps.iter().any(|a| a.id == "cmd"));
-        for id in JETBRAINS_IDE_IDS {
-            assert!(
-                config.apps.iter().any(|a| a.id == *id),
-                "missing JetBrains IDE default entry for {id}"
-            );
-        }
-    }
-
-    /// Backfilling `windows_exe_names` on a user entry that lacks it
-    /// is the mechanism that lights up icon extraction on Windows for
-    /// pre-existing `apps.json` files. The user's other fields must
-    /// stay untouched — we're only filling the gap for a new
-    /// optional field.
-    #[test]
-    fn merge_backfills_windows_exe_names_without_clobbering_user_fields() {
-        let user = AppsConfig {
-            apps: vec![AppEntry {
-                id: "vscode".into(),
-                // Custom name the user has set — must persist.
-                name: "My VSCode".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["my-code".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec![],
-                windows_appx_package: String::new(),
-                open_args: vec!["--user".into(), "{}".into()],
-                needs_terminal: false,
-            }],
-        };
-        let embedded = AppsConfig {
-            apps: vec![AppEntry {
-                id: "vscode".into(),
-                name: "VS Code".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["code".into()],
-                mac_app_names: vec!["Visual Studio Code.app".into()],
-                windows_exe_names: vec!["Code.exe".into()],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-
-        let merged = merge_missing_default_entries(user, embedded);
-        assert_eq!(merged.apps.len(), 1);
-        let entry = &merged.apps[0];
-        // User-customized fields untouched.
-        assert_eq!(entry.name, "My VSCode");
-        assert_eq!(entry.bin_names, vec!["my-code".to_string()]);
-        assert_eq!(
-            entry.open_args,
-            vec!["--user".to_string(), "{}".to_string()]
-        );
-        // New optional fields backfilled from the embedded default.
-        assert_eq!(entry.windows_exe_names, vec!["Code.exe".to_string()]);
-        assert_eq!(
-            entry.mac_app_names,
-            vec!["Visual Studio Code.app".to_string()]
-        );
-    }
-
-    /// If the user has explicitly set `windows_exe_names`, the merge
-    /// must NOT overwrite it. This guards the customize-then-upgrade
-    /// path: an advanced user pointing at a portable / non-default
-    /// install shouldn't have their override silently wiped on
-    /// startup.
-    #[test]
-    fn merge_does_not_overwrite_user_windows_exe_names() {
-        let user = AppsConfig {
-            apps: vec![AppEntry {
-                id: "vscode".into(),
-                name: "VS Code".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["code".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec!["PortableCode.exe".into()],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-        let embedded = AppsConfig {
-            apps: vec![AppEntry {
-                id: "vscode".into(),
-                name: "VS Code".into(),
-                category: AppCategory::Editor,
-                bin_names: vec!["code".into()],
-                mac_app_names: vec![],
-                windows_exe_names: vec!["Code.exe".into()],
-                windows_appx_package: String::new(),
-                open_args: vec!["{}".into()],
-                needs_terminal: false,
-            }],
-        };
-
-        let merged = merge_missing_default_entries(user, embedded);
-        assert_eq!(
-            merged.apps[0].windows_exe_names,
-            vec!["PortableCode.exe".to_string()]
-        );
-    }
-
-    const JETBRAINS_IDE_IDS: &[&str] = &[
-        "intellij",
-        "aqua",
-        "clion",
-        "datagrip",
-        "dataspell",
-        "fleet",
-        "goland",
-        "phpstorm",
-        "pycharm",
-        "rider",
-        "rubymine",
-        "rustrover",
-        "webstorm",
-        "writerside",
-    ];
-
-    #[test]
-    fn embedded_jetbrains_entries_have_cross_platform_detection_metadata() {
-        let config: AppsConfig =
-            serde_json::from_str(DEFAULT_APPS_JSON).expect("default-apps.json must parse");
-
-        for id in JETBRAINS_IDE_IDS {
-            let entry = config
-                .apps
-                .iter()
-                .find(|app| app.id == *id)
-                .unwrap_or_else(|| panic!("missing JetBrains IDE default entry for {id}"));
-            assert_eq!(entry.category, AppCategory::Ide);
-            assert!(
-                !entry.bin_names.is_empty(),
-                "{id} needs Unix/Linux launcher names"
-            );
-            assert!(
-                !entry.mac_app_names.is_empty(),
-                "{id} needs macOS app bundle names"
-            );
-            assert!(
-                !entry.windows_exe_names.is_empty(),
-                "{id} needs Windows exe names"
-            );
-            assert_eq!(entry.open_args, vec!["{}".to_string()]);
-        }
-    }
-
-    #[test]
-    fn detect_finds_jetbrains_ides_from_default_config_when_launchers_are_on_path() {
-        let tmp = tempfile::tempdir().unwrap();
-        for id in JETBRAINS_IDE_IDS {
-            let bin = tmp.path().join(default_bin_name_for_app(id));
-            std::fs::write(&bin, "#!/bin/sh\n").unwrap();
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                std::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o755)).unwrap();
-            }
-        }
-
-        let config: AppsConfig =
-            serde_json::from_str(DEFAULT_APPS_JSON).expect("default-apps.json must parse");
-
-        let detected = detect_with_paths(&config, &[tmp.path().to_path_buf()]);
-        for id in JETBRAINS_IDE_IDS {
-            let app = detected
-                .iter()
-                .find(|app| app.id == *id)
-                .unwrap_or_else(|| panic!("{id} should be detected from the embedded defaults"));
-            assert_eq!(app.category, AppCategory::Ide);
-            assert_eq!(
-                app.detected_path,
-                tmp.path()
-                    .join(default_bin_name_for_app(id))
-                    .to_string_lossy()
-                    .to_string()
-            );
-        }
-    }
-
-    fn default_bin_name_for_app(id: &str) -> &str {
-        if id == "intellij" { "idea" } else { id }
-    }
-
-    #[cfg(target_os = "macos")]
-    #[test]
-    fn jetbrains_toolbox_script_dirs_include_macos_toolbox_location() {
-        let home = Path::new("/Users/example");
-        assert_eq!(
-            jetbrains_toolbox_script_dirs(Some(home)),
-            vec![
-                PathBuf::from("/Users/example")
-                    .join("Library")
-                    .join("Application Support")
-                    .join("JetBrains")
-                    .join("Toolbox")
-                    .join("scripts")
-            ],
-        );
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn jetbrains_toolbox_script_dirs_include_linux_toolbox_location() {
-        let home = Path::new("/home/example");
-        assert_eq!(
-            jetbrains_toolbox_script_dirs(Some(home)),
-            vec![
-                PathBuf::from("/home/example")
-                    .join(".local")
-                    .join("share")
-                    .join("JetBrains")
-                    .join("Toolbox")
-                    .join("scripts")
-            ],
-        );
-    }
-
-    #[cfg(target_os = "windows")]
-    #[test]
-    fn jetbrains_toolbox_script_dirs_include_windows_toolbox_location_without_home() {
-        let dirs = jetbrains_toolbox_script_dirs(None);
-        if let Some(local_data) = dirs::data_local_dir() {
-            assert_eq!(
-                dirs,
-                vec![local_data.join("JetBrains").join("Toolbox").join("scripts")],
-            );
-        } else {
-            assert!(dirs.is_empty());
-        }
-    }
-
-    /// End-to-end Windows-only regression test for icon extraction.
-    /// Targets `cmd.exe` because it's:
-    ///   - always present on every Windows install (no install
-    ///     dependency in CI),
-    ///   - has a real embedded icon resource (not a generic shell
-    ///     fallback), and
-    ///   - is a fixed path (`%WINDIR%\System32\cmd.exe`) so the
-    ///     test exercises both the COM path and the data-URL framing
-    ///     without needing any user config.
-    ///
-    /// Asserts:
-    ///   1. The extractor returns `Some` (the previous shipped
-    ///      version returned None for many inputs because of bad
-    ///      `-Command` quoting),
-    ///   2. The data URL has the expected `data:image/png;base64,`
-    ///      prefix,
-    ///   3. The decoded PNG is at least 64×64. The previous
-    ///      `ExtractAssociatedIcon` path always produced 32×32 — if
-    ///      we regress to that, this assert fails. 64 is a deliberate
-    ///      floor below the IShellItemImageFactory 256 we ask for so
-    ///      Windows builds with shell-image disabled (rare) still
-    ///      pass via the upgraded fallback.
-    #[cfg(target_os = "windows")]
-    #[test]
-    fn extract_windows_icon_data_url_returns_high_res_png_for_cmd_exe() {
-        use base64::Engine as _;
-        use base64::engine::general_purpose::STANDARD;
-
-        let cmd = PathBuf::from(std::env::var_os("WINDIR").unwrap_or_else(|| "C:\\Windows".into()))
-            .join("System32")
-            .join("cmd.exe");
-        assert!(
-            cmd.is_file(),
-            "test precondition: {} must exist on every Windows host",
-            cmd.display()
-        );
-
-        let url = extract_windows_icon_data_url("", &cmd)
-            .expect("icon extraction returned None for cmd.exe — extractor regression");
-
-        let prefix = "data:image/png;base64,";
-        assert!(
-            url.starts_with(prefix),
-            "data URL missing png prefix: {}",
-            &url[..url.len().min(64)]
-        );
-
-        let payload = &url[prefix.len()..];
-        let png = STANDARD
-            .decode(payload)
-            .expect("base64 in data URL must decode");
-
-        // PNG signature: 89 50 4E 47 0D 0A 1A 0A.
-        assert!(png.len() > 200, "PNG body too small: {} bytes", png.len());
-        assert_eq!(
-            &png[..8],
-            b"\x89PNG\r\n\x1a\n",
-            "decoded bytes are not a PNG header"
-        );
-
-        // Parse IHDR (starts at byte 8: 4-byte length + "IHDR" chunk type
-        // + 4-byte width + 4-byte height, all big-endian).
-        let width = u32::from_be_bytes([png[16], png[17], png[18], png[19]]);
-        let height = u32::from_be_bytes([png[20], png[21], png[22], png[23]]);
-        assert!(
-            width >= 64 && height >= 64,
-            "icon resolution regressed to ExtractAssociatedIcon-only output \
-             ({width}x{height}); expected ≥64×64 from IShellItemImageFactory"
-        );
-    }
-
-    /// Bad input must surface as `None` (not panic, not hang). This
-    /// pins the contract that callers can treat the extractor as
-    /// best-effort and fall through to the lucide placeholder icon.
-    #[cfg(target_os = "windows")]
-    #[test]
-    fn extract_windows_icon_returns_none_for_missing_path() {
-        let bogus = PathBuf::from(r"C:\does-not-exist\definitely-not-a-real-binary.exe");
-        let result = extract_windows_icon_data_url("", &bogus);
-        assert!(
-            result.is_none(),
-            "expected None for missing path; got {:?}",
-            result
-        );
-    }
-
-    /// Process-global env-var lock so any future test in this binary
-    /// that mutates `$CLAUDETTE_HOME` (or reads it via
-    /// [`claudette::path::claudette_home`]) doesn't observe another
-    /// test's override mid-run. Mirrors the `env_lock()` pattern in
-    /// `src/path.rs::tests` — cargo runs tests in a thread pool by
-    /// default, so a `--test-threads=1` assumption is not safe.
-    #[cfg(windows)]
-    fn env_lock() -> &'static std::sync::Mutex<()> {
-        static LOCK: std::sync::OnceLock<std::sync::Mutex<()>> = std::sync::OnceLock::new();
-        LOCK.get_or_init(|| std::sync::Mutex::new(()))
-    }
-
-    /// `icon_cache_dir()` must route through `claudette_home` so the
-    /// cache lands under `dev --new` / `dev --clone`'s per-PID sandbox
-    /// instead of the real `~/.claudette/icon-cache/`. A direct
-    /// `dirs::home_dir()` would silently leak across sessions — this
-    /// test pins the override against that regression.
-    ///
-    /// `$CLAUDETTE_HOME` is process-global, so we (a) hold `env_lock()`
-    /// for the duration of the test to serialize against any future
-    /// env-mutating test in this binary, and (b) restore the previous
-    /// value before returning regardless of pass/fail.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_dir_honors_claudette_home_override() {
-        let _guard = env_lock().lock().unwrap_or_else(|p| p.into_inner());
-        let prev = std::env::var_os("CLAUDETTE_HOME");
-        let tmp = tempfile::tempdir().unwrap();
-        // SAFETY: protected by env_lock() so no other test in this
-        // binary is reading CLAUDETTE_HOME (directly or via
-        // claudette_home()) while we mutate it.
-        unsafe {
-            std::env::set_var("CLAUDETTE_HOME", tmp.path());
-        }
-        let dir = icon_cache_dir().expect("home is set, dir should resolve");
-        assert_eq!(dir, tmp.path().join("icon-cache"));
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var("CLAUDETTE_HOME", v),
-                None => std::env::remove_var("CLAUDETTE_HOME"),
-            }
-        }
-    }
-
-    /// Cache key must be stable for the same .exe (same path, size,
-    /// mtime) so a cold launch's extracted icon still hits on the
-    /// next launch — that's the whole point of disk caching.
-    /// `merge_*` and `resolve_windows_*` already pin the upstream
-    /// fields; this pins the downstream key derivation.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_key_stable_for_same_exe() {
-        let tmp = tempfile::tempdir().unwrap();
-        let exe = tmp.path().join("stable.exe");
-        std::fs::write(&exe, b"MZ").unwrap();
-        let k1 = icon_cache_key("", &exe).expect("key");
-        let k2 = icon_cache_key("", &exe).expect("key");
-        assert_eq!(k1, k2);
-    }
-
-    /// Different paths must produce different keys, otherwise two
-    /// apps would clobber each other's cached icons.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_key_differs_per_path() {
-        let tmp = tempfile::tempdir().unwrap();
-        let a = tmp.path().join("a.exe");
-        let b = tmp.path().join("b.exe");
-        std::fs::write(&a, b"MZ").unwrap();
-        std::fs::write(&b, b"MZ").unwrap();
-        let ka = icon_cache_key("", &a).expect("key a");
-        let kb = icon_cache_key("", &b).expect("key b");
-        assert_ne!(ka, kb);
-    }
-
-    /// AppxPackage entries must key off the package name, not the
-    /// .exe path — `wt.exe` and `cmd.exe` both resolve to the same
-    /// system console alias if we only hashed the path, and the UWP
-    /// icon would never get cached separately. Passing different
-    /// packages with the same path must produce different keys.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_key_separates_appx_from_exe() {
-        let tmp = tempfile::tempdir().unwrap();
-        let exe = tmp.path().join("alias.exe");
-        std::fs::write(&exe, b"MZ").unwrap();
-        let k_exe = icon_cache_key("", &exe).expect("exe key");
-        let k_uwp = icon_cache_key("Microsoft.WindowsTerminal", &exe).expect("uwp key");
-        assert_ne!(k_exe, k_uwp);
-    }
-
-    /// Cache miss must surface `None` so `app_icon_data_url` knows
-    /// to fall through to extraction. We exercise the read path with
-    /// a definitely-nonexistent key.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_read_miss_returns_none() {
-        let key = "claudette-test-definitely-not-a-real-key-9q3jklm";
-        // Defensive: clean up if a prior test run left this lying.
-        if let Some(dir) = icon_cache_dir() {
-            let _ = std::fs::remove_file(dir.join(format!("{key}.b64")));
-        }
-        assert!(read_icon_cache(key).is_none());
-    }
-
-    /// Round-trip the cache: write then read must return the exact
-    /// same data URL. Pins the on-disk format so we don't accidentally
-    /// add a header / encoding step that breaks existing entries.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_write_then_read_roundtrips() {
-        let key = format!("claudette-roundtrip-{}", std::process::id());
-        let value = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=";
-        write_icon_cache(&key, value);
-        let got = read_icon_cache(&key).expect("just-written cache must read back");
-        assert_eq!(got, value);
-        // Cleanup so the test is hermetic.
-        if let Some(dir) = icon_cache_dir() {
-            let _ = std::fs::remove_file(dir.join(format!("{key}.b64")));
-        }
-    }
-
-    /// Cache invalidation: changing the file's mtime must produce a
-    /// different cache key so a re-extract triggers next launch.
-    /// This is what makes VS Code's silent auto-update get a fresh
-    /// icon if the vendor ever changes it.
-    #[cfg(windows)]
-    #[test]
-    fn icon_cache_key_changes_on_mtime_change() {
-        let tmp = tempfile::tempdir().unwrap();
-        let exe = tmp.path().join("upgraded.exe");
-        std::fs::write(&exe, b"MZ").unwrap();
-        let k1 = icon_cache_key("", &exe).expect("key1");
-
-        // Bump the file's mtime by writing different content + waiting
-        // for filesystem timestamp granularity (NTFS = 100ns, but
-        // some toolchains coalesce same-second writes; sleep keeps it
-        // honest).
-        std::thread::sleep(std::time::Duration::from_secs(1));
-        std::fs::write(&exe, b"MZ\x90\x90").unwrap();
-        let k2 = icon_cache_key("", &exe).expect("key2");
-        assert_ne!(k1, k2, "cache key did not change after mtime + size delta");
-    }
-
-    /// End-to-end UWP regression: only meaningful on machines where
-    /// Windows Terminal is installed via Appx (default on Windows 11).
-    /// We skip cleanly otherwise so this can run unconditionally on
-    /// any Windows host.
-    ///
-    /// What's pinned: the AppxPackage path returns *something* high-
-    /// res (≥64×64 PNG) for `Microsoft.WindowsTerminal`. If a future
-    /// refactor drops the UWP branch, the icon would silently fall
-    /// back to the generic console glyph at 32×32 and this test
-    /// fails loudly.
-    #[cfg(windows)]
-    #[test]
-    fn extract_windows_icon_resolves_appx_package_logo_for_terminal() {
-        use base64::Engine as _;
-        use base64::engine::general_purpose::STANDARD;
-
-        // Quick gate: skip when WT isn't installed (e.g. Windows
-        // Server hosts without the optional component).
-        let pkg_check = claudette::process::std_command("powershell.exe")
-            .args([
-                "-NoProfile",
-                "-Command",
-                "if (Get-AppxPackage -Name 'Microsoft.WindowsTerminal*' -ErrorAction SilentlyContinue) { 'yes' } else { 'no' }",
-            ])
-            .output();
-        let installed =
-            matches!(pkg_check, Ok(o) if String::from_utf8_lossy(&o.stdout).trim() == "yes");
-        if !installed {
-            eprintln!("skipping: Microsoft.WindowsTerminal not installed");
-            return;
-        }
-
-        // Empty path is fine — the AppxPackage branch fires before we
-        // ever look at line 2.
-        let url = extract_windows_icon_data_url("Microsoft.WindowsTerminal", Path::new(""))
-            .expect("Appx icon extraction returned None for installed package");
-
-        let prefix = "data:image/png;base64,";
-        assert!(url.starts_with(prefix), "missing png prefix: {url}");
-        let png = STANDARD
-            .decode(&url[prefix.len()..])
-            .expect("base64 decodes");
-        assert!(png.len() > 200, "PNG body too small ({} bytes)", png.len());
-        assert_eq!(&png[..8], b"\x89PNG\r\n\x1a\n");
-        let width = u32::from_be_bytes([png[16], png[17], png[18], png[19]]);
-        let height = u32::from_be_bytes([png[20], png[21], png[22], png[23]]);
-        assert!(
-            width >= 64 && height >= 64,
-            "Appx logo regressed below 64x64 ({width}x{height}) — \
-             ExtractAssociatedIcon fallback path may have taken over"
-        );
-    }
-
-    /// On Windows, find_binary should prefer `.exe` over `.cmd` over a
-    /// no-extension shim — mirroring PATHEXT order. This protects
-    /// against the VS Code layout where `bin/code` (a Bash script) and
-    /// `bin/code.cmd` coexist; we want the latter so the icon resolver
-    /// can walk up to `Code.exe` and `Command::new` doesn't try to
-    /// invoke a sh-style script.
-    #[cfg(windows)]
-    #[test]
-    fn find_binary_prefers_exe_then_cmd_then_bare_on_windows() {
-        let tmp = tempfile::tempdir().unwrap();
-        let dir = tmp.path();
-        std::fs::write(dir.join("tool"), "shim").unwrap();
-        std::fs::write(dir.join("tool.cmd"), "@echo off").unwrap();
-        std::fs::write(dir.join("tool.exe"), b"MZ").unwrap();
-
-        let resolved = find_binary("tool", &[dir.to_path_buf()]).unwrap();
-        assert_eq!(resolved, dir.join("tool.exe"));
-
-        std::fs::remove_file(dir.join("tool.exe")).unwrap();
-        let resolved = find_binary("tool", &[dir.to_path_buf()]).unwrap();
-        assert_eq!(resolved, dir.join("tool.cmd"));
-
-        std::fs::remove_file(dir.join("tool.cmd")).unwrap();
-        let resolved = find_binary("tool", &[dir.to_path_buf()]).unwrap();
-        assert_eq!(resolved, dir.join("tool"));
-    }
-
-    /// `resolve_windows_icon_source` must walk up from the detected
-    /// shim to the real `.exe` named in `windows_exe_names`. The
-    /// fixture mirrors VS Code's layout: `bin/code.cmd` one level
-    /// below `Code.exe`.
-    #[cfg(windows)]
-    #[test]
-    fn resolve_windows_icon_walks_up_to_named_exe() {
-        let tmp = tempfile::tempdir().unwrap();
-        let app_root = tmp.path().join("MyApp");
-        let bin_dir = app_root.join("bin");
-        std::fs::create_dir_all(&bin_dir).unwrap();
-        let shim = bin_dir.join("myapp.cmd");
-        std::fs::write(&shim, "@echo off").unwrap();
-        let real_exe = app_root.join("MyApp.exe");
-        std::fs::write(&real_exe, b"MZ").unwrap();
-
-        let entry = AppEntry {
-            id: "myapp".into(),
-            name: "My App".into(),
-            category: AppCategory::Editor,
-            bin_names: vec!["myapp".into()],
-            mac_app_names: vec![],
-            windows_exe_names: vec!["MyApp.exe".into()],
-            windows_appx_package: String::new(),
-            open_args: vec!["{}".into()],
-            needs_terminal: false,
-        };
-
-        assert_eq!(resolve_windows_icon_source(&entry, &shim), real_exe);
-    }
-
-    /// Cursor's `cursor.cmd` lives three directories below `Cursor.exe`
-    /// (`<root>/resources/app/bin/cursor.cmd`). The walk-up cap of five
-    /// must comfortably cover that.
-    #[cfg(windows)]
-    #[test]
-    fn resolve_windows_icon_walks_up_multiple_levels() {
-        let tmp = tempfile::tempdir().unwrap();
-        let app_root = tmp.path().join("cursor");
-        let deep = app_root.join("resources").join("app").join("bin");
-        std::fs::create_dir_all(&deep).unwrap();
-        let shim = deep.join("cursor.cmd");
-        std::fs::write(&shim, "@echo off").unwrap();
-        let real_exe = app_root.join("Cursor.exe");
-        std::fs::write(&real_exe, b"MZ").unwrap();
-
-        let entry = AppEntry {
-            id: "cursor".into(),
-            name: "Cursor".into(),
-            category: AppCategory::Editor,
-            bin_names: vec!["cursor".into()],
-            mac_app_names: vec![],
-            windows_exe_names: vec!["Cursor.exe".into()],
-            windows_appx_package: String::new(),
-            open_args: vec!["{}".into()],
-            needs_terminal: false,
-        };
-
-        assert_eq!(resolve_windows_icon_source(&entry, &shim), real_exe);
-    }
-
-    /// When `windows_exe_names` doesn't match anything in the walk-up,
-    /// fall back to the original detected path so the caller still
-    /// gets a stable input for `ExtractAssociatedIcon`.
-    #[cfg(windows)]
-    #[test]
-    fn resolve_windows_icon_falls_back_to_detected_path() {
-        let tmp = tempfile::tempdir().unwrap();
-        let bin_dir = tmp.path().join("bin");
-        std::fs::create_dir_all(&bin_dir).unwrap();
-        let shim = bin_dir.join("ghost.cmd");
-        std::fs::write(&shim, "@echo off").unwrap();
-
-        let entry = AppEntry {
-            id: "ghost".into(),
-            name: "Ghost".into(),
-            category: AppCategory::Editor,
-            bin_names: vec!["ghost".into()],
-            mac_app_names: vec![],
-            windows_exe_names: vec!["Nonexistent.exe".into()],
-            windows_appx_package: String::new(),
-            open_args: vec!["{}".into()],
-            needs_terminal: false,
-        };
-
-        assert_eq!(resolve_windows_icon_source(&entry, &shim), shim);
-    }
-
-    /// An entry without `windows_exe_names` must skip the walk and
-    /// return the detected path verbatim — that's the existing
-    /// behavior for shimless apps and we don't want to regress it.
-    #[cfg(windows)]
-    #[test]
-    fn resolve_windows_icon_no_walk_without_exe_names() {
-        let tmp = tempfile::tempdir().unwrap();
-        let exe = tmp.path().join("standalone.exe");
-        std::fs::write(&exe, b"MZ").unwrap();
-
-        let entry = AppEntry {
-            id: "standalone".into(),
-            name: "Standalone".into(),
-            category: AppCategory::Terminal,
-            bin_names: vec!["standalone".into()],
-            mac_app_names: vec![],
-            windows_exe_names: vec![],
-            windows_appx_package: String::new(),
-            open_args: vec![],
-            needs_terminal: false,
-        };
-
-        assert_eq!(resolve_windows_icon_source(&entry, &exe), exe);
-    }
-
-    // --- icon-helper cfg-gating regression tests ----------------------
-    //
-    // These exist to catch a class of regression where someone narrows
-    // the `cfg(...)` on `data_url_from_bytes` / `image_data_url_from_file`
-    // and breaks the Linux build, because the Linux icon lookup path
-    // (`find_icon_file_recursive`, `app_icon_data_url`) also calls
-    // `image_data_url_from_file`. That mistake shipped once already
-    // (PR #729's "only consumed by mac_icon_from_app_bundle" claim was
-    // wrong — Linux's XDG icon walk consumes it too) and only surfaced
-    // in nightly because PR CI doesn't compile `claudette-tauri` on
-    // Linux. Every test here is `#[cfg(any(target_os = "macos",
-    // target_os = "linux"))]` or stricter — narrowing either gate will
-    // refuse to compile this module, which is the signal we want.
-
-    /// `data_url_from_bytes` must be reachable from both macOS and Linux
-    /// — both platforms encode discovered icon files as data URLs for
-    /// the workspace opener UI.
-    #[cfg(any(target_os = "macos", target_os = "linux"))]
-    #[test]
-    fn data_url_from_bytes_encodes_payload() {
-        let url = data_url_from_bytes("image/png", b"hello");
-        // base64("hello") == "aGVsbG8="
-        assert_eq!(url, "data:image/png;base64,aGVsbG8=");
-    }
-
-    /// `image_data_url_from_file` must compile and return a data URL
-    /// for known image extensions on both macOS and Linux. The bytes
-    /// don't have to be a real PNG — the helper picks media type from
-    /// the extension and base64-encodes whatever it reads.
-    #[cfg(any(target_os = "macos", target_os = "linux"))]
-    #[test]
-    fn image_data_url_from_file_returns_data_url_for_known_extensions() {
-        let tmp = tempfile::tempdir().unwrap();
-
-        let png = tmp.path().join("icon.png");
-        std::fs::write(&png, b"hello").unwrap();
-        assert_eq!(
-            image_data_url_from_file(&png).as_deref(),
-            Some("data:image/png;base64,aGVsbG8=")
-        );
-
-        let svg = tmp.path().join("icon.svg");
-        std::fs::write(&svg, b"<svg/>").unwrap();
-        assert_eq!(
-            image_data_url_from_file(&svg).as_deref(),
-            Some("data:image/svg+xml;base64,PHN2Zy8+")
-        );
-
-        // Case folding on the extension — covers `.PNG` etc.
-        let upper = tmp.path().join("upper.PNG");
-        std::fs::write(&upper, b"hi").unwrap();
-        assert!(
-            image_data_url_from_file(&upper)
-                .is_some_and(|url| url.starts_with("data:image/png;base64,"))
-        );
-    }
-
-    /// Unknown / non-image extensions must return `None` so the icon
-    /// fallback chain can keep walking. Pinning this guards against a
-    /// future "default to octet-stream" change that would smuggle
-    /// arbitrary files into the UI.
-    #[cfg(any(target_os = "macos", target_os = "linux"))]
-    #[test]
-    fn image_data_url_from_file_rejects_unknown_extension() {
-        let tmp = tempfile::tempdir().unwrap();
-        let txt = tmp.path().join("not-an-icon.txt");
-        std::fs::write(&txt, b"hello").unwrap();
-        assert!(image_data_url_from_file(&txt).is_none());
-
-        let no_ext = tmp.path().join("bare");
-        std::fs::write(&no_ext, b"hello").unwrap();
-        assert!(image_data_url_from_file(&no_ext).is_none());
-    }
-
-    /// End-to-end: the Linux icon lookup walks an XDG-shaped tree and
-    /// must hand the resulting path to `image_data_url_from_file`. This
-    /// is the *exact* call site that broke in nightly — if either the
-    /// recursive walk or the data-URL helper is gated wrong on Linux,
-    /// this test stops compiling on Linux.
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_find_icon_file_recursive_returns_path_consumable_by_data_url_helper() {
-        let tmp = tempfile::tempdir().unwrap();
-        // Mirror the production call shape — `linux_icon_file_from_name`
-        // hands each entry of `linux_icon_roots()` (e.g. `/usr/share/icons`)
-        // to `find_icon_file_recursive`, which expects the
-        // `hicolor/<size>/apps/<icon>` tree to live *inside* that root.
-        let icons_root = tmp.path().join("icons");
-        let icon_dir = icons_root.join("hicolor/16x16/apps");
-        std::fs::create_dir_all(&icon_dir).unwrap();
-        let icon_path = icon_dir.join("myapp.png");
-        std::fs::write(&icon_path, b"hello").unwrap();
-
-        let found = find_icon_file_recursive(&icons_root, "myapp", 6)
-            .expect("recursive walk must find the staged icon");
-        assert_eq!(found, icon_path);
-
-        let data_url =
-            image_data_url_from_file(&found).expect("found path must be encodable by the helper");
-        assert!(data_url.starts_with("data:image/png;base64,"));
-    }
-
-    /// Decoy filename with the same stem but a non-image extension
-    /// must be skipped — `find_icon_file_recursive` filters via
-    /// `image_data_url_from_file`, and that contract is what makes the
-    /// Linux walk safe. Regressing this would let `myapp.txt` shadow
-    /// `myapp.svg` and surface a missing-icon to the UI.
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_find_icon_file_recursive_skips_non_image_extensions() {
-        let tmp = tempfile::tempdir().unwrap();
-        let icons_root = tmp.path().join("icons");
-        // Decoy lives in the shallow `16x16/apps` slot — a hit-by-stem
-        // that the walker must reject because `.txt` isn't an image
-        // type the data-URL helper recognizes.
-        let decoy_dir = icons_root.join("hicolor/16x16/apps");
-        std::fs::create_dir_all(&decoy_dir).unwrap();
-        std::fs::write(decoy_dir.join("myapp.txt"), b"not an icon").unwrap();
-        // Real icon lives one tier deeper in `scalable/apps`. The walker
-        // has to keep going past the decoy to reach it.
-        let real_dir = icons_root.join("hicolor/scalable/apps");
-        std::fs::create_dir_all(&real_dir).unwrap();
-        let real = real_dir.join("myapp.svg");
-        std::fs::write(&real, b"<svg/>").unwrap();
-
-        let found = find_icon_file_recursive(&icons_root, "myapp", 6)
-            .expect("walker must find the .svg even when a .txt decoy exists");
-        assert_eq!(found, real);
-    }
 }
